% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{APTest}
\alias{APTest}
\title{A function for an exhaustive query of Affinity Clustering results when used with
four possible distance metrics: The Pearson & Spearman correlations between
samples, as well as Manhattan & Euclidean distances.  The function will cluster
using Affinity Propagation clustering, with each sample's own similarity being
set as the 0th (min), 25th, 50th (median), 75th, and 100th (max) percentiles
of all data in the similarity matrix.  The fewest clusters should be generated
when the diagonal is set to the minimum.  This function can also be applied
only on 'npcs' Principal Components of the data, and an iteration option is
available to cluster on 2,3,...,npcs principal components using the distance
metric 'iteratemethod' iteratively, to view how the number of PCs impacts
the number of clusters.  data.obj should be transformed with SCTransform}
\usage{
APTest(
  data.obj,
  skipnonpca = FALSE,
  spearman = TRUE,
  pearson = TRUE,
  negmanhattan = TRUE,
  negeuclidean = TRUE,
  PCA = TRUE,
  npcs = 30,
  iterate = FALSE,
  iteratemethod = "manhattan",
  step = 1
)
}
\arguments{
\item{data.obj}{a Seurat object. Cells are columns and features are rows.}

\item{skipnonpca}{set TRUE if you want to skip any calculations not on PCs}

\item{spearman}{test the Spearman distance metric (both non-PCA and PCA)}

\item{pearson}{test the Pearson distance metric (both non-PCA and PCA)}

\item{negmanhattan}{test the Manhattan distance metric (negative) (both non-PCA and PCA)}

\item{negeuclidean}{test the Euclidean distance metric (negative) (both non-PCA and PCA)}

\item{PCA}{set TRUE (default) to perform tests on 'npcs' many PCs of the data}

\item{npcs}{number of PCs to test on.  Iterating also only reaches this many PCs}

\item{iterate}{set TRUE if you want clusters vs. npcs for 'iteratemethod' metric}

\item{iteratemethod}{the distance metric used for PC iteration (must choose only one)}

\item{step}{the step size when iterating over PCs}
}
\value{
A named list of lists. APCluster results
}
\description{
A function for an exhaustive query of Affinity Clustering results when used with
four possible distance metrics: The Pearson & Spearman correlations between
samples, as well as Manhattan & Euclidean distances.  The function will cluster
using Affinity Propagation clustering, with each sample's own similarity being
set as the 0th (min), 25th, 50th (median), 75th, and 100th (max) percentiles
of all data in the similarity matrix.  The fewest clusters should be generated
when the diagonal is set to the minimum.  This function can also be applied
only on 'npcs' Principal Components of the data, and an iteration option is
available to cluster on 2,3,...,npcs principal components using the distance
metric 'iteratemethod' iteratively, to view how the number of PCs impacts
the number of clusters.  data.obj should be transformed with SCTransform
}
