% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{clusterMatrix}
\alias{clusterMatrix}
\title{Takes 2 cluster vectors, which should be named, and makes a matrix with all
incident datapoints.  For example, if data point i is incident on cluster
2 of vector 1 and cluster 5 of vector 2, it will add one count to the cell
(2,5).}
\usage{
clusterMatrix(rowVect, colVect, rowNames = "CLUSTER", colNames = "CLUSTER")
}
\arguments{
\item{rowVect}{the cluster vector to index the rows}

\item{colVect}{the cluster vector to index the cols}

\item{rowNames}{a prefix for each rowVect cluster name (1 -> CLUSTER.1)}

\item{colNames}{a prefix for each colVect cluster name (T-cell -> CLUSTER.T-cell)}
}
\value{
a matrix with all datapoints incident on cluster assignments from rowVect & colVect
}
\description{
Takes 2 cluster vectors, which should be named, and makes a matrix with all
incident datapoints.  For example, if data point i is incident on cluster
2 of vector 1 and cluster 5 of vector 2, it will add one count to the cell
(2,5).
}
