% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{ClusterSimQuantiles}
\alias{ClusterSimQuantiles}
\title{Helper function to run APCluster on a given similarity matrix with diagonals
set to different quantiles of the similarity matrix.  Diagonals are not counted
when calculating quantiles.  Quantiles used include 0th (min), 25th, 50th (med),
75th, 100th(max) percentiles  
Note that on a 5000x5000 similarity matrix, the total size of the output list
is typically around 2.5 Megabytes}
\usage{
ClusterSimQuantiles(sim.mat, name)
}
\arguments{
\item{sim.mat}{A similarity matrix; diagonals will be ignored}

\item{name}{The name of the data whose quantiles are being clustered. Required.}
}
\value{
A named list of APCluster results
}
\description{
Helper function to run APCluster on a given similarity matrix with diagonals
set to different quantiles of the similarity matrix.  Diagonals are not counted
when calculating quantiles.  Quantiles used include 0th (min), 25th, 50th (med),
75th, 100th(max) percentiles  
Note that on a 5000x5000 similarity matrix, the total size of the output list
is typically around 2.5 Megabytes
}
