% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{compareClusters}
\alias{compareClusters}
\title{A function for testing a list of cluster vectors on 5 different metrics.
The 5 metrics used for comparison are the Variational Information (vi.dist),
the Rand Index (RI), the Normalized Information Distance (NID), the Adjusted
Rand Index (ARI), and the Modified Adjusted Rand Index (MARI).  All of these
are metrics defined in Chiquet et al. 2020, and are various computations of
cluster comparisons.  The function returns a table with the metrics for all
vectors compared to the first vector, so the first vector should be the
assignments vector if the computation is intended to compare to true labels.}
\usage{
compareClusters(clustvect.list)
}
\arguments{
\item{clustvect.list}{A named list of cluster vectors, all of same cardinality}
}
\value{
A table, with a row for each vector in the list, and a column for each comparison metric
}
\description{
A function for testing a list of cluster vectors on 5 different metrics.
The 5 metrics used for comparison are the Variational Information (vi.dist),
the Rand Index (RI), the Normalized Information Distance (NID), the Adjusted
Rand Index (ARI), and the Modified Adjusted Rand Index (MARI).  All of these
are metrics defined in Chiquet et al. 2020, and are various computations of
cluster comparisons.  The function returns a table with the metrics for all
vectors compared to the first vector, so the first vector should be the
assignments vector if the computation is intended to compare to true labels.
}
