% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{graphPCvariance}
\alias{graphPCvariance}
\title{A function for generating a 2D plot of the variance explained by each PC of a
Seurat 'data.obj' transformed with RunPCA.  Will plot PCs on X-axis and variance
explained on Y-axis.  Function does not return a result, only prints.Graph 
should be saved by user with ggsave().}
\usage{
graphPCvariance(data.obj)
}
\arguments{
\item{data.obj}{a Seurat object transformed with PCA}
}
\description{
A function for generating a 2D plot of the variance explained by each PC of a
Seurat 'data.obj' transformed with RunPCA.  Will plot PCs on X-axis and variance
explained on Y-axis.  Function does not return a result, only prints.Graph 
should be saved by user with ggsave().
}
