% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{A wrapper function for ggplot2's heatmap and stats::heatmap, with an option
for hierarchical clustering.  This function turns a table of numerical values
into a heatmap, with color intensity based on the value in each cell.}
\usage{
plotHeatmap(table, dodge = 1, hcluster = TRUE)
}
\arguments{
\item{table}{A table with numerical values in each cell}

\item{dodge}{Allow multiple overlaps for the labels}

\item{hcluster}{Perform default hierarchical clustering on heatmap}
}
\description{
A wrapper function for ggplot2's heatmap and stats::heatmap, with an option
for hierarchical clustering.  This function turns a table of numerical values
into a heatmap, with color intensity based on the value in each cell.
}
