% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{SummarizeAPTest_NONPCA_CPM}
\alias{SummarizeAPTest_NONPCA_CPM}
\title{Takes an apresult returned by APTest and makes a data table that summarizes
the number of exemplars produced by each distance metric, using each quantile
of the data, on the CPM matrix returned from Seurat::SCTransform.  
All similarity metrics must be used in this APTest, and there must
be a NON-PCA analysis (skipnonpca = FALSE).}
\usage{
SummarizeAPTest_NONPCA_CPM(apresult)
}
\arguments{
\item{apresult}{an AP clustering result returned by APTest()}
}
\value{
a table that summarizes the NON-PCA CPM experiment
}
\description{
Takes an apresult returned by APTest and makes a data table that summarizes
the number of exemplars produced by each distance metric, using each quantile
of the data, on the CPM matrix returned from Seurat::SCTransform.  
All similarity metrics must be used in this APTest, and there must
be a NON-PCA analysis (skipnonpca = FALSE).
}
